Refactor the app structures
