+++ src/reducers/navReducer.js
@@ -1,4 +1,4 @@
-import AppNavigator from '../appNavigator';
+import AppNavigator from '../navigations/appNavigator';
 
 const navReducer = (state, action) => {
   const newState = AppNavigator.router.getStateForAction(action, state);
